requirements.txt
django
django-jazzmin
django-import-export
pandas
numpy
nltk
scikit-learn
